# api/message_bus/openapi.yaml
openapi: 3.0.3
info:
  title: CasaOS Message Bus API
  version: v2
  description: |
    CasaOS Message Bus accepts events and actions from various sources and delivers them to subscribers.

    For issues, please visit the [GitHub repository](https://github.com/ReCasaOS/CasaOS-MessageBus).
servers:
  - url: /v2/message_bus
...
security:
  - access_token: []
paths:
...
  /event/{source_id}:
    get:
      summary: Subscribe to events by source ID (WebSocket)
      description: |
        Subscribe to event by `source_id` via WebSocket.

        > `names` can be specified in the query string to subscribe to specific event types.
      operationId: subscribeEventWS
      security:
        - access_token: []
        - access_token_query: []
      tags:
        - WebSocket Endpoints
      parameters:
        - $ref: "#/components/parameters/SourceID"
        - $ref: "#/components/parameters/EventNames"
      responses:
        "101":
          description: |
            The connection will be upgraded to a WebSocket connection and the client will receive events as they are published.
...
  /action/{source_id}:
    get:
      summary: Subscribe to actions by source ID (WebSocket)
      description: |
        Subscribe to actions by `source_id` via WebSocket.

        > `names` can be specified in the query string to subscribe to specific action types.
      operationId: subscribeActionWS
      security:
        - access_token: []
        - access_token_query: []
      tags:
        - WebSocket Endpoints
      parameters:
        - $ref: "#/components/parameters/SourceID"
        - $ref: "#/components/parameters/ActionNames"
      responses:
        "101":
          description: |
            The connection will be upgraded to a WebSocket connection and the client will receive actions as they are triggered.
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
      description: |
        A CasaOS user JWT. This box's own services send `Internal <secret>` instead, from loopback, with the per-boot secret the gateway writes to `/var/run/casaos/internal.secret`; requests on the unix socket need neither.
    access_token_query:
      type: apiKey
      in: query
      name: token
      description: |
        The same user JWT, in the query string, because browsers cannot set headers on a WebSocket. Read on the subscription routes only (the WebSocket and socket.io endpoints) and ignored everywhere else.
  parameters:
    SourceID:
      name: source_id
      in: path
      required: true
      schema:
        type: string
        example: "local-storage"
...
    EventNames:
      name: names
      in: query
      allowEmptyValue: false
      required: false
      schema:
        type: array
        items:
          type: string
      example: "local-storage:disk:added,local-storage:disk:removed"
...
    ActionNames:
      name: names
      in: query
      allowEmptyValue: false
      required: false
      schema:
        type: array
        items:
          type: string
      example: "local-storage:partition:format,local-storage:partition:mount"
...